requirements.txt
langchain_community
langchain_google_genai
langchain
streamlit
python-dotenv
chromadb
pypdf2
sentence_transformers